add automatic start

+++ app/src/main/java/com/example/xealnfc/User.kt
@@ -2,6 +2,6 @@ package com.example.xealnfc
 
 data class User(
     var name: String?,
-    var remainingAmount: Int?,
-    var previousAmount: Int?
+    var remainingAmount: Double?,
+    var previousAmount: Double?
 )
